+++ vite-plugin-ssr/node/runtime/page-files/setup.ts
@@ -1,7 +1,7 @@
 import { setPageFilesAsync } from '../../../shared/getPageFiles'
 import { assert, debugGlob } from '../utils'
 import { getGlobalContext } from '../globalContext'
-import { virtualModuleIdPageFilesServer } from '../plugin/plugins/generateImportGlobs/virtualModuleIdPageFiles'
+import { virtualModuleIdPageFilesServer } from '../../plugin/plugins/generateImportGlobs/virtualModuleIdPageFiles'
 
 setPageFilesAsync(getPageFilesExports)
